% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utls.R
\name{coverageRate}
\alias{coverageRate}
\title{Calculates the coverage rate of estimated clusters}
\usage{
coverageRate(clusters.true, clusters.result)
}
\arguments{
\item{clusters.true}{A list of vectors, each vector contains the elements in the true cluster.}

\item{clusters.result}{A list of vectors, each vector contains the elements in the estimated cluster.}
}
\value{
The coverage rate of estimated clusters.
}
\description{
Calculates the coverage rate of estimated clusters
}
