% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBEMA.R
\name{dBEMA}
\alias{dBEMA}
\title{Estimate the number of latent factors in B.}
\usage{
dBEMA(Stand.B, N, alpha = 0.2, n.avg = 10)
}
\arguments{
\item{Stand.B}{Coefficient matrix.}

\item{N}{Sample size.}

\item{alpha}{Quantile of bulk eigenvalues.}

\item{n.avg}{Number of simulations to estimate the distribution of bulk eigenvalues.}
}
\value{
The number of latent factors.
}
\description{
For this to work properly, the input coefficient matrix 'Stand.B' should be properly standardized.
}
