% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_G.R
\name{fitMixGaussian}
\alias{fitMixGaussian}
\title{Estimate the parameters in Gaussian mixture model for G.hat.}
\usage{
fitMixGaussian(
  G,
  init_sigma,
  init_pi,
  fix.sigma1 = TRUE,
  threshold.converage = 1e-05,
  threshold.fsr = 0.05
)
}
\arguments{
\item{G}{G.hat. It will be standarded to have norm sqrt(S) if it's not.}

\item{init_sigma}{A numeric vector of length 2, containing the initial standard deviations of the Gaussian components}

\item{init_pi}{A numeric vector of length 2, containing the initial weights of the Gaussian components}

\item{fix.sigma1}{If the standard deviation of the null should be fixed. Default is TRUE.}

\item{threshold.fsr}{The threshold of local false sign rate to conclude significant SNPs. Default is 0.05.}
}
\value{
A list contatining three objects. parameters: the posterior probabilities and variances of Gaussian componients. significantSNPs: A logic vector indicating if the SNPs are significant. lfsr: the local false sign rates of all SNPs.
}
\description{
Estimate the parameters in Gaussian mixture model for G.hat.
}
\examples{
fit <- fitMixGaussian(G=rnorm(1000),init_sigma=c(1,1.5),init_pi=c(0.9,0.1))
}
