% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utls.R
\name{graphicalCluster}
\alias{graphicalCluster}
\title{Combining the subpathways sharing parameters with large probability to form clusters}
\usage{
graphicalCluster(
  w,
  method = "max",
  threshold,
  include.singleton = F,
  p.singleton = NULL
)
}
\arguments{
\item{w}{The similarity matrix. It's rows and columns should be named.}

\item{method}{The method used to measure the similarity between one element and a group of elements. Can take values of "average", "min" and "max". Default is "max".}

\item{include.singleton}{Should singleton subpathways be included in estimated clusters. Default is False.}

\item{p.singleton}{If 'include.singleton' is True and 'p.singleton' is provided, only subpathways that do not share parameter with other pathways with large probability are considered as true singletons.}
}
\value{
A list of vectors, each vector contains the elements in the estimated cluster.
}
\description{
Combining the subpathways sharing parameters with large probability to form clusters
}
