% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_sampler.R
\name{HDP_Gibbs}
\alias{HDP_Gibbs}
\title{Gibbs sampler used to estimate the posterior distribution of parameters in HDP model for metabolite effects}
\usage{
HDP_Gibbs(data, Iter = 1500, record = 501)
}
\arguments{
\item{data}{A list contains the input. It can be generated by function 'prepareData'.}

\item{Iter}{A integer specifying the total number of iterations.}

\item{record}{A integer specifying the iteration at which the function starts recording Gibbs samples.}
}
\value{
A list containing the Gibbs samples of each factor.
"Mean": a number of Gibbs samples by number of subpathways matrix, whose entries represent the posterior means of subpathways;
"SD": a number of Gibbs samples by number of subpathways matrix, whose entries represent the posterior sd of subpathways;
"Outlier": a number of Gibbs samples by number of metabolites matrix, whose entries indicate if the metabolite is an outlier (1) or not (0);
"L.post.mean": a number of Gibbs samples by number of metabolites matrix, whose entries records the posterior mean of each metabolite;
"L.post.sd": a number of Gibbs samples by number of metabolites matrix, whose entries records the posterior sds of each metabolite ;
"lfsr": a number of Gibbs samples by number of metabolites matrix, whose entries records the local false sign rate of each metabolite ;
"W": a number of subpathways by number of subpathways matrix, with it (i,j)th entry represents the probability that subpathway i and j share the same parameter;
"p.singleton": a length number of subpathways vector, recording the probability that each subpathway do not share parameter with other subpathways;
"P.post.theta": a length number of Iter vector, recording the log joint posterior probability of parameters at each iteration.
}
\description{
Gibbs sampler used to estimate the posterior distribution of parameters in HDP model for metabolite effects
}
