% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utls.R
\name{inclusionRate}
\alias{inclusionRate}
\title{Calculates the inclusion rate of estimated clusters}
\usage{
inclusionRate(clusters.true, clusters.result)
}
\arguments{
\item{clusters.true}{A list of vectors, each vector contains the elements in the true cluster.}

\item{clusters.result}{A list of vectors,  each vector contains the elements in the estimated cluster.}
}
\value{
The inclusion rate of estimated clusters.
}
\description{
Calculates the inclusion rate of estimated clusters
}
