% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simuFactorModel}
\alias{simuFactorModel}
\title{Simulate metabolite GWAS data from the factor model}
\usage{
simuFactorModel(
  IDs = NULL,
  J = 5,
  K = 10,
  N = 1000,
  S = 10000,
  G.model = c("gaussian", "laplace"),
  p.l.spike = 0.05,
  p.G.nonzero = 0.005,
  p.out = 0.01,
  p.Delta.0 = 0.2,
  common.spikes = NULL,
  regenerate.L = FALSE,
  regenerate.paras = NULL,
  fix.L.scaling = FALSE,
  L.scaling = NULL
)
}
\arguments{
\item{IDs}{A data.frame object specifying the pathway structure. It will be randomly generated if unspecified.}

\item{J}{Number of superpathways used to generate IDs. Default is 5.}

\item{K}{Number of latent factors. Default is 10.}

\item{N}{Sample size. Default is 1000.}

\item{S}{Number of genetic variants. Default is 1e4.}

\item{G.model}{Distribution of non-zero genetic effects. Can take values of either 'gaussian' or 'laplace'.}

\item{p.l.spike}{Probability of generating spike pathways.}

\item{p.G.nonzero}{Probability of generating non-zero entries in G. Default is 0.005.}

\item{p.out}{Probability of generating outliers.}

\item{p.Delta.0}{Probability of non-genetically regulated metabolites in Delta. Default is 0.2.}

\item{common.spikes}{A vector of subpathway names. All entries of L belonging to those subpathway are set as 0. Default is NULL.}

\item{regenerate.paras}{A list containing parameters to regenerate L. Default is NULL.}

\item{fix.L.scaling}{Default is FALSE. The function will automatically scale L to make the signal strength compariable to noise level. If it's set as TRUE, 'L.scaling' must also be specified.}

\item{L.scaling}{User specificed scalar to scale L.}
}
\value{
A list containg matrices in mtGWAS factor analysis model.
}
\description{
Simulate metabolite GWAS data from the factor model
}
