% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simuPathway}
\alias{simuPathway}
\title{Simulate a super-pathway and sub-pathway file}
\usage{
simuPathway(J)
}
\arguments{
\item{J}{Number of super-pathways to simulate}
}
\value{
A data frame of three coloumns, namely the superpathway, subpathway and metabolite indices.
}
\description{
Simulate a super-pathway and sub-pathway file
}
\details{
Only number of super-pathways is required to be specified, the number of subpathways and total number of metabolites will be randomly generated.
}
\examples{
IDs <- simuPathway(5)
}
