% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_functions.R
\name{prepareData}
\alias{prepareData}
\title{Generate the data format used by HDP_Gibbs.}
\usage{
prepareData(B.hat, N, K, IDs)
}
\arguments{
\item{B.hat}{The standardized coefficient matrix.}

\item{N}{Sample size.}

\item{K}{Number of latent factors to estimate.}

\item{IDs}{Metabolite pathway information.}
}
\value{
A list containing estimated L and variances of estimates.
}
\description{
Generate the data format used by HDP_Gibbs.
}
